chore(server-components): move parent class to components/server/index.js
