+++ components/contributor-spotlight/index.js
@@ -2,7 +2,7 @@ import { html } from "lit";
 
 import { Section } from "../../components/content/index.js";
 import { PageLayout } from "../../components/page-layout/index.js";
-import { ServerComponent } from "../../components/server.js";
+import { ServerComponent } from "../server/index.js";
 
 export class ContributorSpotlight extends ServerComponent {
   /**
