+++ components/not-found/index.js
@@ -1,7 +1,7 @@
 import { html } from "lit";
 
 import { PageLayout } from "../page-layout/index.js";
-import { ServerComponent } from "../server.js";
+import { ServerComponent } from "../server/index.js";
 
 export class NotFound extends ServerComponent {
   /**
