+++ components/observatory-results/index.js
@@ -2,7 +2,7 @@ import { html } from "lit";
 
 import { Feedback } from "../observatory-landing/feedback.js";
 import { PageLayout } from "../page-layout/index.js";
-import { ServerComponent } from "../server.js";
+import { ServerComponent } from "../server/index.js";
 
 export class ObservatoryResults extends ServerComponent {
   /**
