+++ components/search/index.js
@@ -1,7 +1,7 @@
 import { html } from "lit";
 
 import { PageLayout } from "../../components/page-layout/index.js";
-import { ServerComponent } from "../server.js";
+import { ServerComponent } from "../server/index.js";
 
 export class Search extends ServerComponent {
   /**
